Move libvfio-user processing to different threads
